Minor Changes. Fix Statistics and erase some comments. Working OK

+++ DiceApp_JDBC/src/main/java/com/mdkGame/DiceApp_JDBC/Domain/Stats.java
@@ -15,7 +15,7 @@ public class Stats {
 		this.qtGames = qtGames;
 		this.qtIsWin = qtIsWin;
 		if(qtGames != 0) {
-			this.avgIsWin = this.getQtIsWin() / this.getQtGames();
+			this.avgIsWin = (float) this.getQtIsWin() / this.getQtGames();
 		}else {
 			this.avgIsWin = (float) 0.0;
 		}
